# Micro-optimizations of the geopotential computations