--- a/physics/geopotential.hpp
+++ b/physics/geopotential.hpp
@@ -39,8 +39,7 @@ class Geopotential {
   GeneralSphericalHarmonicsAcceleration(
       Instant const& t,
       Displacement<Frame> const& r,
-      Square<Length> const& r²,
-      Exponentiation<Length, -3> const& one_over_r³) const;
+      Square<Length> const& r²) const;
 
  private:
   // The frame of the surface of the celestial.