Benchmark (note that the Fortran code has become slower due to the benchmark being now fair):
```
Run on (4 X 3310 MHz CPU s)
09/30/18 16:32:13
--------------------------------------------------------------------
Benchmark                             Time           CPU Iterations
--------------------------------------------------------------------
BM_ComputeGeopotentialCpp/2      193212 ns     193208 ns      13807
BM_ComputeGeopotentialCpp/3      249143 ns     248923 ns      11218
BM_ComputeGeopotentialCpp/5      380883 ns     380278 ns       7179
BM_ComputeGeopotentialCpp/10    1065107 ns    1061599 ns       2601
BM_ComputeGeopotentialF90/2      186733 ns     186696 ns      14957
BM_ComputeGeopotentialF90/3      237447 ns     235970 ns      11966
BM_ComputeGeopotentialF90/5      320568 ns     319412 ns       8547
BM_ComputeGeopotentialF90/10     826904 ns     821305 ns       3324
```
C++ fit: `165282 - 2179.71 n + 9177.35 n²`
Fortran fit: `175488 - 4669.92 n + 6919.05 n²`

So the inner loop is about 30% more expensive in C++.